Move PythonRequirement to src/python/pants/python.
