+++ src/python/pants/backend/python/BUILD
@@ -7,13 +7,13 @@ python_library(
   dependencies = [
     ':pants_requirement',
     ':python_artifact',
-    ':python_requirement',
     ':python_requirements',
     'src/python/pants/backend/python/rules',
     'src/python/pants/backend/python/targets',
     'src/python/pants/backend/python/tasks',
     'src/python/pants/build_graph',
     'src/python/pants/goal:task_registrar',
+    'src/python/pants/python',
   ],
   tags = {"partially_type_checked"},
 )
@@ -24,8 +24,8 @@ python_library(
   dependencies = [
     ':interpreter_cache',
     ':python_artifact',
-    ':python_requirement',
     ':python_requirements',
+    'src/python/pants/python',
   ],
   tags = {"partially_type_checked"},
 )
@@ -58,11 +58,11 @@ python_library(
   name = 'pants_requirement',
   sources = ['pants_requirement.py'],
   dependencies = [
-    'src/python/pants/backend/python:python_requirement',
     'src/python/pants/backend/python/targets',
     'src/python/pants/base:build_environment',
     'src/python/pants/base:exceptions',
     'src/python/pants/build_graph',
+    'src/python/pants/python',
     'src/python/pants/util:meta',
   ],
   tags = {"partially_type_checked"},
@@ -80,6 +80,10 @@ python_library(
 python_library(
   name = 'python_requirement',
   sources = ['python_requirement.py'],
+  dependencies = [
+    'src/python/pants/base:deprecated',
+    'src/python/pants/python'
+  ],
   tags = {"partially_type_checked"},
 )
