+++ src/python/pants/backend/python/rules/BUILD
@@ -58,6 +58,7 @@ python_tests(
     'src/python/pants/engine:rules',
     'src/python/pants/engine:selectors',
     'src/python/pants/engine/legacy:structs',
+    'src/python/pants/python',
     'src/python/pants/rules/core',
     'src/python/pants/source',
     'src/python/pants/testutil:interpreter_selection_utils',
