+++ src/python/pants/backend/python/targets/BUILD
@@ -6,11 +6,11 @@ python_library(
     '3rdparty/python:pex',
     '3rdparty/python/twitter/commons:twitter.common.collections',
     'src/python/pants/backend/python:python_artifact',
-    'src/python/pants/backend/python:python_requirement',
     'src/python/pants/base:exceptions',
     'src/python/pants/base:payload',
     'src/python/pants/base:payload_field',
     'src/python/pants/build_graph',
+    'src/python/pants/python',
   ],
   tags = {"partially_type_checked"},
 )
