+++ src/python/pants/backend/python/tasks/BUILD
@@ -12,7 +12,6 @@ python_library(
     'src/python/pants/backend/native/targets',
     'src/python/pants/backend/native/tasks',
     'src/python/pants/backend/python:interpreter_cache',
-    'src/python/pants/backend/python:python_requirement',
     'src/python/pants/backend/python/lint/isort',
     'src/python/pants/backend/python/rules',  # For setup_py_util.py.
     'src/python/pants/backend/python/subsystems',
