+++ src/python/pants/python/BUILD
@@ -5,6 +5,7 @@
 python_library(
   dependencies=[
     '3rdparty/python:pex',
+    '3rdparty/python:setuptools',
     'src/python/pants/base:build_environment',
     'src/python/pants/base:exceptions',
     'src/python/pants/build_graph',
