+++ src/python/pants/python/pex_build_util.py
@@ -12,12 +12,12 @@ from pex.resolver import resolve
 from pex.util import DistributionHelper
 from twitter.common.collections import OrderedSet
 
-from pants.backend.python.python_requirement import PythonRequirement
 from pants.base.build_environment import get_buildroot
 from pants.base.exceptions import TaskError
 from pants.build_graph.files import Files
 from pants.build_graph.target import Target
 from pants.python.python_repos import PythonRepos
+from pants.python.python_requirement import PythonRequirement
 from pants.python.python_setup import PythonSetup
 from pants.subsystem.subsystem import Subsystem
 from pants.util.collections import assert_single_element
