+++ tests/python/pants_test/backend/project_info/tasks/BUILD
@@ -19,8 +19,8 @@ python_tests(
     'src/python/pants/backend/jvm/targets:jvm',
     'src/python/pants/backend/project_info/tasks:dependencies',
     'src/python/pants/backend/python/targets',
-    'src/python/pants/backend/python:python_requirement',
     'src/python/pants/build_graph',
+    'src/python/pants/python',
     'src/python/pants/testutil:task_test_base',
   ],
   tags = {"partially_type_checked"},
