+++ tests/python/pants_test/backend/python/BUILD
@@ -5,10 +5,10 @@ python_tests(
   name='pants_requirement',
   sources=['test_pants_requirement.py'],
   dependencies=[
-    'src/python/pants/backend/python:python_requirement',
     'src/python/pants/backend/python:plugin',
     'src/python/pants/backend/python/targets',
     'src/python/pants/base:build_environment',
+    'src/python/pants/python',
     'src/python/pants/testutil:test_base',
   ],
   tags = {"partially_type_checked"},
@@ -19,9 +19,9 @@ python_tests(
   sources=['test_python_requirement_list.py'],
   dependencies=[
     'src/python/pants/backend/python/targets',
-    'src/python/pants/backend/python:python_requirement',
     'src/python/pants/base:exceptions',
     'src/python/pants/build_graph',
+    'src/python/pants/python',
     'src/python/pants/testutil:test_base'
   ],
   tags = {"partially_type_checked"},
