+++ tests/python/pants_test/backend/python/targets/BUILD
@@ -4,7 +4,7 @@
 python_tests(
   dependencies=[
     'src/python/pants/backend/python/targets',
-    'src/python/pants/backend/python:python_requirement',
+    'src/python/pants/python',
     'src/python/pants/util:collections',
     'src/python/pants/testutil:test_base',
   ],
