+++ tests/python/pants_test/backend/python/test_python_requirement_list.py
@@ -3,10 +3,10 @@
 
 from textwrap import dedent
 
-from pants.backend.python.python_requirement import PythonRequirement
 from pants.backend.python.targets.python_requirement_library import PythonRequirementLibrary
 from pants.base.exceptions import TargetDefinitionException
 from pants.build_graph.build_file_aliases import BuildFileAliases
+from pants.python.python_requirement import PythonRequirement
 from pants.testutil.test_base import TestBase
